Fixing some edge cases

+++ lib/churn_vs_complexity/complexity/pmd/files_calculator.rb
@@ -8,9 +8,12 @@ module ChurnVsComplexity
           def folder_based? = false
 
           def calculate(files:)
+            return Parser.empty_result if files.empty?
+
             cache_path = PMD.resolve_cache_path
             files_arg = files.map { |file| "-d #{file}" }.join(' ')
-            output = `pmd check #{files_arg} -R #{PMD.resolve_ruleset_path} -f json -t #{CONCURRENCY} --cache #{cache_path} 2>/dev/null`
+            command = "pmd check #{files_arg} -R #{PMD.resolve_ruleset_path} -f json -t #{CONCURRENCY} --cache #{cache_path} 2>/dev/null"
+            output = `#{command}`
             File.delete(cache_path)
 
             Parser.new.parse(output)
